requirements.txt
pillow ~= 10.4.0
